Checkpoint update and Prefabs
Updated the Template scene with the new camera system and updated prefabs.
Updated the state of the prefabs to ensure they are drag and drop out of the Template scene.
Duplicated Halen Room to keep an original version for reference once changes from another branch are merged.
Updated the LevelManager, Checkpoints, and player respawning to avoid using FindObjectOfType as much, plus some other improvements to code, such as pointless member references.