currently available sounds implemented
Removed Millivolt.Sound namespace (lots of things use it, might as well put it in the main namespace).
Added error handling to the SFXController.
Changed the footstep animation event script into the PlayerAnimationEventController in case any other sounds need to be tied to animations.
Added logic to objects that had sounds available to use.