Player rotation fix
thanks to adam, fixed the player gravity rotation.
Moved all rotation logic and the collider to the player model object off of the PlayerController object.
Updated all Event Objects to check for the PlayerModel script instead of the PlayerController object.
Moved the EventObject script into the correct folder.
Removed the defunct sound controller object off of the Player prefab.
GameManager ChangeGravity method calls the OnGravityChange for the player objects.