Fixed quitting to menu breaking the scene loading
Issue was a missing GameManager reference since it stays in destroy on load after the game is opened, removing the reference when re-opening the scene. Now calls a method on the MainMenu script that calls the original GameManager method through its static instance on a delay so that the CRT effect still plays (it was getting eaten during the scene loading).