Started work on Options menu and getting the Teleporter UI to work
Added the appropriate sections to the options menu ui and set up swapping between the submenus in options.
Need to now hook up sliders on the menu to the actual settings they change and make it accessable in the pause menu.
Started working on updating the checkpoint teleporter ui menu to have it work with the existing UI pieces already in the project