Dissolving screw! and sfx stuff
Made a script that makes an object dissolve before it is destroyed using the dissolve shader and some magic.
Created a physics material for the screw.
Imported the screw model that is one mesh.
Changed some scripts to use TryGetComponent.
Changed some of the interaction logic to allow checkpoints to appear in the InteractionUI.
Added some basic stuff to the SFXController for sound types for volume stuff later.