Failed camera fix + Blending added + misc
Moved deprecated scripts into a deprecated folder in Assets/Camera/.
Added camera sensitivity changing logic into the camera controller.
Removed deprecated inputs from the input manager and PlayerInteraction enum.
Moved CameraController into the camera folder.
Tried to fix the camera not pausing when timescale is 0, but turns out that was our solution to the jitter, so we're doing a band-aid fix where we just set the camera speed to 0 whenever the game pauses and setting it back to the proper speed on un-pause.