Fixed lighting, completed options menu functionality
The lighting is now fixed and no longer being not there when restarting the level.
LevelData will now hold the name of the scene that it is storing the data of level for. Switched the main scene to have the correct data object.
Options menu is now hooked up to process controller and keyboard inputs correctly. Made Music labeled audio play in 2D sound so that you can hear it (specifically on the main menu).
Fixed a problem with the UIController that caused it to leave the pausemenu when in the options menu from pause.
Fixed a typo in the camera controller.
Assigned pause menu options button to go to the pause menu instead of restarting the level.
TODO: fix player camera putting you in the ground when flipping upside down, and add Yash's cutscenes