Changed camera again
Fixed the issue with the high look camera by simply removing it.
One camera now changes between two targets to give the different view.
FOV changing script is available but deemed as not necessary, kept in script in case of feedback.
Fixed issue with the GameManager not having a default currentSceneName value.