Added first pass of the emotions being able to change
Added the PlayerEmotions script which will be added to the player model object.
To use:
FIRST make sure all the emotion materials are places into the array.
SECOND set the index value in the PlayerEmotions component to what element the face material is in the material list on the playermodel
THIRD call the functions in PlayerEmotions to change the face, the parameter on the event is how long you want the face to last for.