Level Start fix
Hid the isLoading bool in the inspector on the GameManager.
Fixed interactionUI not being disabled and Player being able to move on level start and restart level. It was PickupObjects in the scene that the player is not holding, as when they were destroyed, even if the Player is not holding it, it still called DropObject.