MORE beta rush hour fixes
Stopped the respawn particle from fritzing out on some checkpoints (reason unsure).
REALLY made it so that checkpoints are only interactable in the editor, meaning the teleporter won't work in builds.
Added a respawn button for the screws in room 2.
REALLY fixed the screw destroying/respawning issue.
Turns out you could interact with things while the game is paused, so I fixed that too.