Pickup system now uses rigidbody force
Changed what makes the held pickup track the held object offset to be AddForce so it can still use collisions and avoid going through walls.
Uses some drag-changing stuff to ensure the behaviour of the screw is good regardless of whether it is being held.
Also added the gravity arrow PNG.