gravity indicator UI
Replaced all WaitForEndOfFrame()s with null in Coroutines because it was actual intended behaviour.
Created the Gravity Indicator Canvas which shows whether the player is on the floor or ceiling.
Added the Canvas to the Player Scene, where it is attached to the PlayerModel and is changed when the PlayerModel rotates with gravity.
Set up fun easings and colours for the Gravity Indicator UI.