Can now make a list and spawn off the list
Credits can now be compiled into an easy list.
Now need to set up the player.
TODO: Add a delay time before spawning, a life time, and then I should be able to give it off to the artists/designers to make the credits perfect.
Also set up a credits list at the end, and a skip button.