Update Main Menu player model
Figured out why the player materials keep detecting changes on them and applied the fix in PlayerEmotion, plus updated all instances that have the PlayerEmotion component.
Moved another player model into a deprecated folder.
Updated the prefab for the MainMenuPlayerModel to use the new player model and tested the animations.
Requires the emotion animation events to be re-set up with the new player model.